Rename file to reflect variable name change
