+++ shared/src/store/action-reducers/exercise.ts
@@ -11,8 +11,8 @@ import { Patient } from '../../models';
 import { getStatus } from '../../models/utils';
 import type { ExerciseState } from '../../state';
 import type { Mutable } from '../../utils';
-import type { ElementTypePluralMap } from '../../utils/type-state-selector-map';
-import { elementTypePluralMap } from '../../utils/type-state-selector-map';
+import type { ElementTypePluralMap } from '../../utils/element-type-plural-map';
+import { elementTypePluralMap } from '../../utils/element-type-plural-map';
 import { IsValue } from '../../utils/validators';
 import type { Action, ActionReducer } from '../action-reducer';
 import { ReducerError } from '../reducer-error';
