+++ shared/src/utils/element-type-plural-map.ts
